Use use-react geolocation library
