+++ package.json
@@ -20,6 +20,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-split-pane": "^0.1.92",
+    "react-use": "^17.2.1",
     "styled-components": "^5.2.1",
     "typescript": "^4.1.5",
     "use-places-autocomplete": "^1.9.1",
